@@ -5,6 +5,7 @@
   "homepage": "/react-win7-build",
   "dependencies": {
     "@babel/core": "7.1.0",
+    "@loadable/component": "^5.7.0",
     "@svgr/webpack": "2.4.1",
     "axios": "^0.18.0",
     "babel-core": "7.0.0-bridge.0",
@@ -30,12 +31,15 @@
     "eslint-plugin-react": "7.11.1",
     "file-loader": "2.0.0",
     "fs-extra": "7.0.0",
+    "github-markdown-css": "^3.0.1",
+    "highlight.js": "^9.15.6",
     "html-webpack-plugin": "4.0.0-alpha.2",
     "identity-obj-proxy": "3.0.0",
     "immutable": "^4.0.0-rc.12",
     "jest": "23.6.0",
     "jest-pnp-resolver": "1.0.1",
     "jest-resolve": "23.6.0",
+    "marked": "^0.6.1",
     "mini-css-extract-plugin": "0.4.3",
     "node-sass": "^4.9.4",
     "optimize-css-assets-webpack-plugin": "5.0.1",
@@ -44,13 +48,14 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.0.6",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.6.0",
+    "react": "^16.8.6",
     "react-app-polyfill": "^0.1.3",
     "react-dev-utils": "^6.0.5",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.8.6",
     "react-hammerjs": "^1.0.1",
     "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
+    "react-syntax-highlighter": "^10.2.1",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "resolve": "1.8.1",
@@ -118,7 +123,7 @@
     ]
   },
   "devDependencies": {
-    "yarn": "^1.12.3",
-    "webpack-dev-server": ">=3.1.11"
+    "webpack-dev-server": ">=3.1.11",
+    "yarn": "^1.12.3"
   }
 }
