@@ -1,19 +1,19 @@
-import {HashRouter, Route, Switch} from 'react-router-dom'
-import React, {Component} from 'react'
-
+import { BrowserRouter, Route, Switch } from 'react-router-dom'
+import React, { Component } from 'react'
+import lazy from '../public/lazy'
 //所有页面
 import Index from '../view/index/index'
-
-export default class Router extends Component{
-    render(){
+const ministore = lazy('mymd/mymd', {})
+export default class extends Component {
+    render() {
         return (
-            <HashRouter>
+            <BrowserRouter>
                 <Switch>
                     {/*<Route redirect='/' path='/*'/>*/}
                     <Route exact path='/' component={Index} />
-
+                    <Route exact path='/ministore' component={ministore} />
                 </Switch>
-            </HashRouter>)
+            </BrowserRouter>)
     }
 }
 
