@@ -3,7 +3,7 @@ import style from './window.scss'
 import classnames from 'classnames'
 import Html from './html/html'
 import MenuArea from '../../../components/menuArea/menuArea'
-import {connect} from 'react-redux'
+import { connect } from 'react-redux'
 import {
     focusApp,
     changeMoveState,
@@ -17,13 +17,25 @@ import {
 class Window extends React.Component {
     constructor(props) {
         super(props)
-        this.windowStartMove = this.windowStartMove.bind(this)
-        this.toggleMaxWindow = this.toggleMaxWindow.bind(this)
-        this.closeWindow = this.closeWindow.bind(this)
-        this.windowEndMove = this.windowEndMove.bind(this)
-        this.windowMoving = this.windowMoving.bind(this)
-        this.clickMax = this.clickMax.bind(this)
-        this.focusWindow = this.focusWindow.bind(this)
+        let bodyStyle = {};
+        //内容控制
+        let content = '';
+        switch (this.props.appItem.detail.type) {
+            case 'txt':
+                content = <p>{this.props.appItem.detail.content}</p>
+                break;
+            case 'html':
+                bodyStyle.overflow = 'hidden'
+                content = <Html url={this.props.appItem.detail.url} />
+                break;
+            case 'component':
+                let App = this.props.appItem.detail.component;
+                content = <App />;
+                break;
+            default:
+                content = <p>{this.props.appItem.detail.content}</p>
+                break;
+        }
         let w = document.documentElement.offsetWidth * 80 / 100,
             h = document.documentElement.offsetHeight * 80 / 100,
             x = document.documentElement.offsetWidth * 10 / 100,
@@ -46,29 +58,26 @@ class Window extends React.Component {
                 h
             }
         }
-
-        this.customState = {}
-    }
-
-    componentWillMount() {
-        // this.props.saveComponent(this, this.props.appItem.detail.appId);
+        this.customState = {
+            bodyStyle,
+            content,
+        }
     }
 
     // 开始移动
-    windowStartMove(e) {
+    windowStartMove = (e) => {
         if (e.button == 0) {
             this.customState.originX = e.pageX;
             this.customState.originY = e.pageY;
             this.customState.startX = this.state.position.x;
             this.customState.startY = this.state.position.y;
-            this.props.changeMovingState(true);
             window.addEventListener('mousemove', this.windowMoving);
             window.addEventListener('mouseup', this.windowEndMove);
         }
     }
 
     //移动过程中
-    windowMoving(e) {
+    windowMoving = (e) => {
         //如果已经放大
         if (this.props.appItem.max) {
             let position = this.state.position;
@@ -86,46 +95,50 @@ class Window extends React.Component {
             })
             this.customState.startX = this.state.position.x;
             this.customState.startY = this.state.position.y;
-        }
-        let position = this.state.position;
-        let currentX = e.pageX;
-        if (e.pageX > document.documentElement.offsetWidth) {
-            currentX = document.documentElement.offsetWidth;
-        }
-        if (e.pageX < 0) {
-            currentX = 0;
-        }
-        let currentY = e.pageY;
-        if (currentY > document.documentElement.offsetHeight - 40) {
-            currentY = document.documentElement.offsetHeight - 40;
-        }
-        if (currentY <= 0) {
-            currentY = 0;
-            this.props.WillMaxCtrl(true);
         } else {
-            this.props.WillMaxCtrl(false);
+            let position = this.state.position;
+            let currentX = e.pageX;
+            if (e.pageX > document.documentElement.offsetWidth) {
+                currentX = document.documentElement.offsetWidth;
+            }
+            if (e.pageX < 0) {
+                currentX = 0;
+            }
+            let currentY = e.pageY;
+            if (currentY > document.documentElement.offsetHeight - 40) {
+                currentY = document.documentElement.offsetHeight - 40;
+            }
+            if (currentY <= 0) {
+                currentY = 0;
+                this.props.WillMaxCtrl(true);
+            } else {
+                this.props.WillMaxCtrl(false);
+            }
+            position.x = this.customState.startX + currentX - this.customState.originX;
+            position.y = this.customState.startY + currentY - this.customState.originY;
+            this.setState({
+                position
+            })
         }
-        position.x = this.customState.startX + currentX - this.customState.originX;
-        position.y = this.customState.startY + currentY - this.customState.originY;
-        this.setState({
-            position
-        })
+
     }
 
     //结束移动
-    windowEndMove() {
-        //如果将放大
-        if (this.props.willMax) {
-            this.props.WillMaxCtrl(false);
-            this.props.maxWindow();
+    windowEndMove = () => {
+        if (this.props.moving) {
+            //如果将放大
+            if (this.props.willMax) {
+                this.props.WillMaxCtrl(false);
+                this.props.maxWindow();
+            }
+            this.props.changeMovingState(false);
         }
-        this.props.changeMovingState(false);
         window.removeEventListener('mousemove', this.windowMoving);
         window.removeEventListener('mouseup', this.windowEndMove);
     }
 
     //点击放大缩小
-    clickMax() {
+    clickMax = () => {
         let position = this.state.position;
         if (position.y < 0) {
             position.y = 50;
@@ -136,7 +149,7 @@ class Window extends React.Component {
         this.toggleMaxWindow();
     }
 
-    toggleMaxWindow() {
+    toggleMaxWindow = () => {
         if (this.props.appItem.max) {
             this.props.cancelMaxWindow();
         } else {
@@ -144,12 +157,14 @@ class Window extends React.Component {
         }
     }
 
-    closeWindow() {
+    closeWindow = () => {
         this.props.closeWindow(this.props.appItem.detail.appId)
     }
 
-    focusWindow() {
-        this.props.focusWindow(this.props.appItem.detail.appId)
+    focusWindow = () => {
+        if (!this.props.isFocus) {
+            this.props.focusWindow(this.props.appItem.detail.appId)
+        }
     }
 
     render() {
@@ -181,25 +196,7 @@ class Window extends React.Component {
         let maskStyle = {
             display: this.props.moving ? "block" : "none"
         }
-        let bodyStyle = {};
-        //内容控制
-        let content = '';
-        switch (this.props.appItem.detail.type) {
-            case 'txt':
-                content = <p>{this.props.appItem.detail.content}</p>
-                break;
-            case 'html':
-                bodyStyle.overflow = 'hidden'
-                content = <Html url={this.props.appItem.detail.url}/>
-                break;
-            case 'component':
-                let App = this.props.appItem.detail.component;
-                content = <App/>;
-                break;
-            default:
-                content = <p>{this.props.appItem.detail.content}</p>
-                break;
-        }
+
         return (
             <div onMouseDown={this.focusWindow} style={winStyle} className={classnames({
                 [style['my_win']]: true,
@@ -208,18 +205,18 @@ class Window extends React.Component {
             })}>
                 <div className={style['win-cell']}>
                     <div onDoubleClick={this.toggleMaxWindow} onClick={this.windowEndMove}
-                         onMouseDown={this.windowStartMove} className={style['win-head']}>
-                        <MenuArea menu={this.menu}/>
+                        onMouseDown={this.windowStartMove} className={style['win-head']}>
+                        <MenuArea menu={this.menu} />
                         <span
-                            style={{backgroundImage: `url(${this.props.appItem.detail.img})`}}>{this.props.appItem.detail.name}</span>
+                            style={{ backgroundImage: `url(${this.props.appItem.detail.img})` }}>{this.props.appItem.detail.name}</span>
                         <div className={style['win-btn']}>
-                            <span onClick={this.props.hideWindow} style={{background: '#8ec831'}}></span>
-                            <span onClick={this.clickMax} style={{background: '#ffd348'}}></span>
-                            <span onClick={this.closeWindow} style={{background: '#ed4646'}}></span>
+                            <span onClick={this.props.hideWindow} style={{ background: '#8ec831' }}></span>
+                            <span onClick={this.clickMax} style={{ background: '#ffd348' }}></span>
+                            <span onClick={this.closeWindow} style={{ background: '#ed4646' }}></span>
                         </div>
                     </div>
-                    <div style={bodyStyle} className={style['win-body']}>
-                        {content}
+                    <div style={this.customState.bodyStyle} className={style['win-body']}>
+                        {this.customState.content}
                         <div style={maskStyle} className={style['win-mask']}></div>
                     </div>
 
@@ -231,13 +228,14 @@ class Window extends React.Component {
 }
 
 export default connect(
-    state => {
+    (state, { appItem }) => {
         return {
+            isFocus: state.runList.find(item => item.appId === appItem.appId).focus,
             moving: state.moving
         }
     },
-    (dispatch, {appItem}) => {
-        const {appId} = appItem
+    (dispatch, { appItem }) => {
+        const { appId } = appItem
         return {
             maxWindow() {
                 dispatch(maxApp(appId))
